style: Replace invalid SPDX license expression
When doing a `yarn e n`, I get a warning to use a valid SPX license expression relating to line of package.json being "GPL".

According to the [documentation](https://spdx.org/licenses), "GPL" is an invalid license. We should probably change this to use the SPDX version to correctly reflect licensing.